requirements.txt
aws-cdk-lib==2.141.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-ecs
aws-cdk.aws-ecr
aws-cdk.aws-route53
aws-cdk.aws-certificatemanager
aws-cdk.aws-ecs-patterns